Update NotificationCreator to use MediaControllerCompat.
I'm not sure if this is a good idea or not, but it's a thing
that I did. The upside is that NupService doesn't need to
pass a bunch of state. I guess it was good practice for a
future update to make NupActivity use the controller.

Also make the media session's queue just use indexes instead
of song IDs to identify items. I'm not completely sure that
this is correct, but the docs say that multiple items
shouldn't have the same ID, which is something that I was
violating before.