Add adaptive launcher icon.
I'm not sure that I'm happy with how this looks, but it at
least gets rid of the ugly white background. When minimizing
the app, there's an ugly point in the animation where the
bright gradient in the background is visible below the
foreground.

I haven't deleted the old bitmap launcher icon since it's
still used for Android Auto.