Automatically unpause when 'replace' button is clicked.
Try to improve things when the user replaces a
previously-restored playlist (which would've resulted in us
already being paused). See #23.

This is still a bit weird, because it means that if the user
explicitly pauses using the button in a browse activity, the
'replace' button will unpause while the 'append' and
'insert' buttons won't.