Add Cloud Build test config and scripts.
Add a cloudbuild/ directory with a test.yaml config.

I'm holding off on building a Docker container since those
are a pain to manage, and I'm bad enough at Docker that I
suspect that my container may take longer to download than
this code currently takes to install the dependencies.