Clear playlist before media session preset search.
This makes Android Auto display a 'Getting your
selection...' message while the search happens in the
background. Fixes #31.

Also make MediaSessionManager report errors for empty
searches and for playback errors.

Unrelatedly, update searchServer() to only send the
'minRating' parameter when it's greater than 0.
Otherwise, it seems to conflict with 'unrated'.