Minor improvements to playback position reporting.
Make the position-reporting task in the Player class
reschedule itself every time that it runs instead of relying
on a recurring schedule (which seems that it could drift
over time).

Also remove code in NupActivity that tried to limit the
frequency of updates of time-related labels, since the
Player class already limits updates to once per second.