Use black text when drawing over light cover images.
Compute the average color of the bottom-left region of
bitmaps when loading them, and update NupActivity to make
overlaid text be either white or black based on the color's
luminosity to improve legibility. This is far from perfect,
but it seems better than always using white text.

I initially experimented with using the Palette class to
compute bitmaps' dominant colors, but it didn't produce
reliable results (and sometimes oddly seemed to just return
the fallback color).