Don't include text in app bar action items.
Stop setting 'withText' on menu items; it looks terrible due
to there being almost zero padding between the icons and
text when the items are displayed in the app bar on a
tablet.

Also move the settings item earlier in the main menu and
explicitly set other items to never appear in the app bar.